modularize components
